+++ packages/modal/src/Component.mobile.tsx
@@ -1,15 +1,18 @@
 import React, { cloneElement, forwardRef, isValidElement } from 'react';
+import cn from 'classnames';
 import { useMedia } from '@alfalab/hooks';
 
-import { Modal, ModalProps } from './Component';
+import { BaseModal, BaseModalProps } from '@alfalab/core-components-base-modal';
+
 import { HeaderMobile } from './components/header/Component.mobile';
 import { ContentMobile } from './components/content/Component.mobile';
 import { FooterMobile } from './components/footer/Component.mobile';
 import { Closer } from './components/closer/Component';
 
-import transitions from './transitions/content.module.css';
+import styles from './mobile.module.css';
+import transitions from './transitions.module.css';
 
-export type ModalMobileProps = Omit<ModalProps, 'fullscreen' | 'size'> & {
+export type ModalMobileProps = BaseModalProps & {
     /**
      * Управление наличием закрывающего крестика
      * @default false
@@ -18,7 +21,7 @@ export type ModalMobileProps = Omit<ModalProps, 'fullscreen' | 'size'> & {
 };
 
 const ModalMobileComponent = forwardRef<HTMLDivElement, ModalMobileProps>(
-    ({ children, ...restProps }, ref) => {
+    ({ children, className, ...restProps }, ref) => {
         const [size] = useMedia(
             [
                 ['s', '(max-width: 375px)'],
@@ -28,22 +31,24 @@ const ModalMobileComponent = forwardRef<HTMLDivElement, ModalMobileProps>(
         );
 
         return (
-            <Modal
+            <BaseModal
                 {...restProps}
                 ref={ref}
                 transitionProps={{
                     classNames: transitions,
                     ...restProps.transitionProps,
                 }}
-                hideBackdrop={true}
-                fullscreen={true}
+                className={cn(className, styles.component)}
+                backdropProps={{
+                    invisible: true,
+                }}
             >
                 {React.Children.map(children, child =>
                     isValidElement(child)
-                        ? cloneElement(child, { size: child.props.size || size })
+                        ? cloneElement(child, { size: child.props.size || size, fullscreen: true })
                         : child,
                 )}
-            </Modal>
+            </BaseModal>
         );
     },
 );