+++ packages/modal/src/components/content/Component.mobile.tsx
@@ -1,16 +1,10 @@
 import React, { FC } from 'react';
-import cn from 'classnames';
 import { Content, ContentProps } from './Component';
 
 import styles from './mobile.module.css';
 
-export type ContentMobileProps = ContentProps & {
-    /**
-     * Размер
-     */
-    size?: 's' | 'm';
-};
+export type ContentMobileProps = ContentProps;
 
-export const ContentMobile: FC<ContentMobileProps> = ({ size, className, ...restProps }) => (
-    <Content className={cn(className, size && styles[size])} {...restProps} />
+export const ContentMobile: FC<ContentMobileProps> = props => (
+    <Content styles={styles} {...props} />
 );