+++ packages/modal/src/components/footer/Component.mobile.tsx
@@ -1,16 +1,8 @@
 import React, { FC } from 'react';
-import cn from 'classnames';
 import { Footer, FooterProps } from './Component';
 
 import styles from './mobile.module.css';
 
-export type FooterMobileProps = FooterProps & {
-    /**
-     * Размер
-     */
-    size?: 's' | 'm';
-};
+export type FooterMobileProps = FooterProps;
 
-export const FooterMobile: FC<FooterMobileProps> = ({ size, className, ...restProps }) => (
-    <Footer className={cn(className, size && styles[size])} {...restProps} />
-);
+export const FooterMobile: FC<FooterMobileProps> = props => <Footer styles={styles} {...props} />;