+++ packages/amount/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.17](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-amount@1.0.16...@alfalab/core-components-amount@1.0.17) (2021-04-09)
+
+**Note:** Version bump only for package @alfalab/core-components-amount
+
+
+
+
+
 ## [1.0.16](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-amount@1.0.15...@alfalab/core-components-amount@1.0.16) (2021-04-01)
 
 **Note:** Version bump only for package @alfalab/core-components-amount